state estimation added
to enable state estimation provided by matpower, more measurement of
lines and transformers are added and used.

The state estimation program is revisied a little to enhance its
convergency (using deltX to judge the convergence, instead of deltY)

see the initialConfig.m
find the seEnable option, which will switch the state estimation on or
off.

However, there are more test requred for this new function, with various
configuration of state estimation.